third_party/WebKit/LayoutTests/http/tests/inspector/elements/styles/edit-css-with-source-url-expected.txt

Tests file system project mappings.


Running: testEditingRulesInElementsPanelDoesNotAddSourceURLToOriginalFile
Adding file system.
Adding file system mapping.
Dumping uiSourceCode content:
#inspected {
    color: red;
}

Loading stylesheet with sourceURL:
Stylesheet loaded.
Dumping matched rules:
[expanded] 
element.style { ()

[expanded] 
#inspected { (foo.css:1 -> foo.css:1:13)
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;

Editing styles from elements panel:
Styles edited.
Dumping matched rules:
[expanded] 
element.style { ()

[expanded] 
#inspected { (foo.css:1 -> foo.css:1:13)
    color: green;
    : ;

[expanded] 
div { (user agent stylesheet)
    display: block;

Dumping uiSourceCode content:
#inspected {
    color: green;
}

All done.
